[transfer-engine] Ensure group notification is published with the correct preview body.
MNotificationGroup::publish() sets its x-nemo-preview-body field
using the preview body of the last added notification group. This
means if the group is published before the notification, the
preview body used will be that of the last published notification
group instead of the current one.

If the group is published after the notification instead, the
notification will ensure the group properties are set correctly
before the group is published.